separate phantom slots from vanilla slots
