+++ src/main/java/com/cleanroommc/modularui/api/widget/IVanillaSlot.java
@@ -5,11 +5,12 @@ import net.minecraft.inventory.Slot;
 /**
  * Marks a {@link IWidget}, that this is a vanilla item slot.
  */
-@FunctionalInterface
 public interface IVanillaSlot {
 
     /**
      * @return the item slot of this widget
      */
     Slot getVanillaSlot();
+
+    boolean handleAsVanillaSlot();
 }
