+++ src/main/java/com/cleanroommc/modularui/screen/ClientScreenHandler.java
@@ -19,7 +19,7 @@ import com.cleanroommc.modularui.utils.Animator;
 import com.cleanroommc.modularui.utils.Color;
 import com.cleanroommc.modularui.utils.FpsCounter;
 import com.cleanroommc.modularui.widget.sizer.Area;
-import com.cleanroommc.modularui.widgets.ItemSlot;
+import com.cleanroommc.modularui.widgets.slot.ItemSlot;
 import com.cleanroommc.modularui.widgets.RichTextWidget;
 import com.cleanroommc.modularui.widgets.slot.ModularSlot;
 import com.cleanroommc.modularui.widgets.slot.SlotGroup;
@@ -369,7 +369,7 @@ public class ClientScreenHandler {
 
         acc.setHoveredSlot(null);
         IGuiElement hovered = muiScreen.getContext().getHovered();
-        if (hovered instanceof IVanillaSlot vanillaSlot) {
+        if (hovered instanceof IVanillaSlot vanillaSlot && vanillaSlot.handleAsVanillaSlot()) {
             acc.setHoveredSlot(vanillaSlot.getVanillaSlot());
         }
