+++ src/main/java/com/cleanroommc/modularui/widgets/SlotGroupWidget.java
@@ -4,6 +4,8 @@ import com.cleanroommc.modularui.api.widget.ISynced;
 import com.cleanroommc.modularui.api.widget.IWidget;
 import com.cleanroommc.modularui.widget.ParentWidget;
 
+import com.cleanroommc.modularui.widgets.slot.ItemSlot;
+
 import it.unimi.dsi.fastutil.chars.Char2IntMap;
 import it.unimi.dsi.fastutil.chars.Char2IntOpenHashMap;
 import it.unimi.dsi.fastutil.chars.Char2ObjectMap;
